backend/drf/requirements.txt
aiofile==3.3.3
aiofiles==0.6.0
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
caio==0.6.3
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
Django==3.1.3
django-cors-headers==3.5.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.4
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.2.0
